# pishdad-core/backend/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Pishdad Core API
  description: |-
    قرارداد هسته بک‌اند — این سند فقط مسیرهای همین هسته را توصیف می‌کند.
    همه پیام‌ها فارسی. احراز هویت با Bearer (Sanctum) + کوکی httpOnly `auth_token`.
    افزونه‌ها مسیرهای خودشان را از `manifest.json` اعلام می‌کنند و `PluginRouteTable`
    آن‌ها را در catch-allِ `v1/p/{slug}` ثبت می‌کند؛ سندِ هر افزونه جداست.
     نسخه ۱.۱۱: داده‌های محتوایی و مدیریتی نصب بین همه مدیران مشترک است؛ user_id فقط سازنده/حسابرسی است. پروفایل، نشست‌ها و داده‌های مالی شخصی از این قاعده مستثنا هستند.
  version: 1.11.0
servers:
  - url: http://localhost:8080
    description: Docker local (nginx → php-fpm)
security: []
paths:
...
  /api/v1/admin/managers:
    get:
      summary: لیست مدیران — سوپرادمین همیشه مخفی (تسک ۵.۱)
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: لیست مدیران }
        '403': { description: نیازمند دسترسی users.view }
    post:
      summary: ساخت مدیر + تخصیص نقش spatie (رمز قوی) (تسک ۵.۱)
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, email, password]
              properties:
                name: { type: string }
                email: { type: string, format: email }
                password: { type: string, description: 'حداقل ۱۰ نویسه + بزرگ/کوچک/عدد/ویژه' }
                role: { type: string, enum: [owner, admin, editor, viewer] }
      responses:
        '201': { description: مدیر ساخته شد }
        '403': { description: نیازمند دسترسی users.edit }
        '422': { $ref: '#/components/responses/ValidationError' }
  /api/v1/admin/managers/{id}:
    get:
      summary: نمایش مدیر (سوپرادمین = 403) (تسک ۵.۱)
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        '200': { description: مدیر }
        '403': { description: دسترسی مجاز نیست }
    put:
      summary: ویرایش مدیر + تغییر نقش (تسک ۵.۱)
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        '200': { description: مدیر به‌روزرسانی شد }
        '403': { description: دسترسی مجاز نیست }
    delete:
      summary: حذف مدیر (خود + سوپرادمین ممنوع) (تسک ۵.۱)
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        '200': { description: مدیر حذف شد }
        '403': { description: دسترسی مجاز نیست }
        '422': { description: حذف حساب خود مجاز نیست }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Sanctum
...
  responses:
...
    ValidationError:
      description: خطای اعتبارسنجی ورودی
      content:
        application/json:
          schema:
            type: object
            properties:
              message: { type: string }
              errors: { type: object }
...